+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # [2.3.x]
 - optimize function
+- solve problem of keyboard shelter toast
 
 # [2.1.x]
 - reconstruction of the underlying logic